Restructures Layout folder
