+++ src/components/Layout/Footer/Slider/Slider.styled.jsx
