+++ src/components/Layout/Footer/Slider/index.jsx
