+++ src/components/Layout/Header/CartIcon/CartIcon.styled.jsx
